+++ src/splat/segtypes/linker_entry.py
@@ -4,7 +4,7 @@ from functools import lru_cache
 from pathlib import Path
 from typing import Dict, List, OrderedDict, Set, Tuple, Union, Optional
 
-from ..util import options
+from ..util import options, log
 
 from .segment import Segment
 from ..util.symbols import to_cname
@@ -236,7 +236,7 @@ class LinkerWriter:
 
         section_entries: OrderedDict[str, List[LinkerEntry]] = OrderedDict()
         for l in segment.section_order:
-            if l in options.opts.ld_section_labels:
+            if l in options.opts.section_order:
                 section_entries[l] = []
 
         # Add all entries to section_entries
@@ -244,7 +244,7 @@ class LinkerWriter:
         for entry in entries:
             if entry.section_order_type in section_entries:
                 # Search for the very first section type
-                # This is required in case the very first entry is a type that's not listed on ld_section_labels (like linker_offset) because it would be dropped
+                # This is required in case the very first entry is a type that's not listed on section_order (like linker_offset) because it would be dropped
                 prev_entry = entry.section_order_type
                 break
 
@@ -252,9 +252,17 @@ class LinkerWriter:
         any_noload = False
         for entry in entries:
             if entry.section_order_type in section_entries:
+                if entry.section_order_type not in section_entries:
+                    log.error(
+                        f"\nError on linker script generation: section '{entry.section_order_type}' not found.\n  Make sure the section '{entry.section_order_type}' is listed on 'section_order' of the yaml options."
+                    )
                 section_entries[entry.section_order_type].append(entry)
             elif prev_entry is not None:
-                # If this section is not present in section_order or ld_section_labels then pretend it is part of the last seen section, mainly for handling linker_offset
+                # If this section is not present in section_order or section_order then pretend it is part of the last seen section, mainly for handling linker_offset
+                if prev_entry not in section_entries:
+                    log.error(
+                        f"\nError on linker script generation: section '{prev_entry}' not found.\n  Make sure the section '{prev_entry}' is listed on 'section_order' of the yaml options."
+                    )
                 section_entries[prev_entry].append(entry)
             any_load = any_load or not entry.noload
             any_noload = any_noload or entry.noload
@@ -285,14 +293,14 @@ class LinkerWriter:
 
         # To keep track which sections has been started
         started_sections: Dict[str, bool] = {
-            l: False for l in options.opts.ld_section_labels
+            l: False for l in options.opts.section_order
         }
 
         # Find where sections are last seen
         last_seen_sections: Dict[LinkerEntry, str] = {}
         for entry in reversed(entries):
             if (
-                entry.section_order_type in options.opts.ld_section_labels
+                entry.section_order_type in options.opts.section_order
                 and entry.section_order_type not in last_seen_sections.values()
             ):
                 last_seen_sections[entry] = entry.section_order_type
@@ -362,7 +370,7 @@ class LinkerWriter:
             self._begin_segment(segment, seg_name, noload=False, is_first=is_first)
 
             for l in segment.section_order:
-                if l not in options.opts.ld_section_labels:
+                if l not in options.opts.section_order:
                     continue
                 if l == ".bss":
                     continue
@@ -412,16 +420,24 @@ class LinkerWriter:
 
         section_entries: OrderedDict[str, List[LinkerEntry]] = OrderedDict()
         for l in segment.section_order:
-            if l in options.opts.ld_section_labels:
+            if l in options.opts.section_order:
                 section_entries[l] = []
 
         # Add all entries to section_entries
         prev_entry = None
         for entry in entries:
             if entry.section_order_type in section_entries:
+                if entry.section_order_type not in section_entries:
+                    log.error(
+                        f"\nError on linker script generation: section '{entry.section_order_type}' not found.\n  Make sure the section '{entry.section_order_type}' is listed on 'section_order' of the yaml options."
+                    )
                 section_entries[entry.section_order_type].append(entry)
             elif prev_entry is not None:
-                # If this section is not present in section_order or ld_section_labels then pretend it is part of the last seen section, mainly for handling linker_offset
+                # If this section is not present in section_order or section_order then pretend it is part of the last seen section, mainly for handling linker_offset
+                if prev_entry not in section_entries:
+                    log.error(
+                        f"\nError on linker script generation: section '{prev_entry}' not found.\n  Make sure the section '{prev_entry}' is listed on 'section_order' of the yaml options."
+                    )
                 section_entries[prev_entry].append(entry)
             prev_entry = entry.section_order_type